SmartResa/backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.2
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
Django==5.2.1
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
msgpack==1.1.0
pillow==11.2.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qrcode==8.2
redis==6.0.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2